backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pytest
httpx  # Required for FastAPI TestClient
